+++ Migrations/Mysql/Version20220615074317.php
@@ -18,6 +18,11 @@ final class Version20220615074317 extends AbstractMigration
     {
         $this->abortIf($this->connection->getDatabasePlatform()->getName() !== 'mysql', 'Migration can only be executed safely on \'mysql\'.');
 
+        $this->skipIf(
+            !\array_key_exists('neos_contentrepository_domain_model_workspace', $this->sm->listTableNames()),
+            'These migrations are not supported in Neos 9'
+        );
+
         // Remove all disconnected workspace details entities or the foreign table constraint will fail
         $this->addSql('DELETE FROM shel_neos_workspacemodule_domain_model_workspacedetails WHERE workspacename NOT IN (SELECT name FROM neos_contentrepository_domain_model_workspace)');
 
@@ -33,9 +38,15 @@ final class Version20220615074317 extends AbstractMigration
     {
         $this->abortIf($this->connection->getDatabasePlatform()->getName() !== 'mysql', 'Migration can only be executed safely on \'mysql\'.');
 
-        $this->addSql('DROP TABLE shel_neos_workspacemodule_domain_model_workspace_1536f_acl_join');
-        $this->addSql('ALTER TABLE shel_neos_workspacemodule_domain_model_workspacedetails DROP FOREIGN KEY FK_923CCEA8D940019');
-        $this->addSql('DROP INDEX flow_identity_shel_neos_workspacemodule_domain_model_work_e2fe7 ON shel_neos_workspacemodule_domain_model_workspacedetails');
-        $this->addSql('ALTER TABLE shel_neos_workspacemodule_domain_model_workspacedetails ADD workspacename VARCHAR(255) CHARACTER SET utf8mb4 NOT NULL COLLATE `utf8mb4_unicode_ci`, DROP workspace');
+        $tables = $this->sm->listTableNames();
+        if (\array_key_exists('shel_neos_workspacemodule_domain_model_workspace_1536f_acl_join', $tables)) {
+            $this->addSql('DROP TABLE shel_neos_workspacemodule_domain_model_workspace_1536f_acl_join');
+        }
+
+        if (\array_key_exists('shel_neos_workspacemodule_domain_model_workspacedetails', $tables)) {
+            $this->addSql('ALTER TABLE shel_neos_workspacemodule_domain_model_workspacedetails DROP FOREIGN KEY FK_923CCEA8D940019');
+            $this->addSql('DROP INDEX flow_identity_shel_neos_workspacemodule_domain_model_work_e2fe7 ON shel_neos_workspacemodule_domain_model_workspacedetails');
+            $this->addSql('ALTER TABLE shel_neos_workspacemodule_domain_model_workspacedetails ADD workspacename VARCHAR(255) CHARACTER SET utf8mb4 NOT NULL COLLATE `utf8mb4_unicode_ci`, DROP workspace');
+        }
     }
 }
